Initial version
This is still experimental and not tested. But it does contain all the
necessary nodetypes and rendering in fusion.
Still many loose edges, e.g.: Do we want our own Collections or do we
just hook into Neos Collections? Could be tricky, and if not we need to
offer a migration.